Text-To-Speech/Azure-Talking-Avatar/requirements.txt
streamlit
python-dotenv
requests